move search_space to configs
